# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowMind AI - Enterprise Workflow Agent",
    "description": "Evidence-grounded, human-governed, auditable workflow AI agent for Customer Complaint Escalation.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/users/me": {
      "get": {
        "tags": [
          "Workflow & Governance"
        ],
        "summary": "Get currently authenticated user identity",
        "description": "Inspect the server-resolved identity of the current token.",
        "operationId": "get_me_api_users_me_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-User-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workflow/investigate": {
      "post": {
        "tags": [
          "Workflow & Governance"
        ],
        "summary": "Start a customer complaint investigation",
        "description": "Initiates the investigation pipeline:\n  retrieval -> reasoning -> recommendation -> (abstention | auto-execute | pending approval)",
        "operationId": "investigate_complaint_api_workflow_investigate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-User-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintInvestigationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflow/{workflow_id}": {
      "get": {
        "tags": [
          "Workflow & Governance"
        ],
        "summary": "Get workflow state and details",
        "description": "Retrieve the current state, evidence citations, and recommendation for a workflow.",
        "operationId": "get_workflow_details_api_workflow__workflow_id__get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflow/{workflow_id}/decision": {
      "post": {
        "tags": [
          "Workflow & Governance"
        ],
        "summary": "Submit human approval decision (RBAC-gated)",
        "description": "Submit a human governance decision (APPROVE, REJECT, or MODIFY).\nCRITICAL: Approver role is resolved strictly from the server session, preventing forgery.",
        "operationId": "submit_workflow_decision_api_workflow__workflow_id__decision_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "X-User-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalSubmission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflow/{workflow_id}/audit": {
      "get": {
        "tags": [
          "Workflow & Governance"
        ],
        "summary": "Get complete tamper-evident audit trail",
        "description": "Retrieve the finalized, state-dependent audit trail for a workflow.",
        "operationId": "get_workflow_audit_api_workflow__workflow_id__audit_get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditExportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflow/{workflow_id}/audit/verify": {
      "get": {
        "tags": [
          "Workflow & Governance"
        ],
        "summary": "Verify SHA-256 tamper-evident audit hash chain",
        "description": "Re-derive every event hash server-side from stored content and confirm chain integrity.\n\nThis endpoint is entirely independent of the write path: it calls verify_event_chain()\nfrom audit/chain.py which recomputes SHA-256 from scratch for each event and checks both\n(a) content integrity (recomputed hash matches stored block_hash) and\n(b) chain linkage (stored parent_hash matches previous event's block_hash).\n\nReturns a real boolean result indicating whether the chain is intact, plus the index and\nevent_id of the first failing event if the chain is broken.",
        "operationId": "verify_workflow_audit_chain_api_workflow__workflow_id__audit_verify_get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComplaintInvestigationRequest": {
        "properties": {
          "customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          },
          "issue_summary": {
            "type": "string",
            "title": "Issue Summary"
          },
          "target_ticket_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Ticket Id"
          },
          "user_role": {
            "type": "string",
            "title": "User Role",
            "default": "support_agent"
          },
          "top_k_tickets": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K Tickets",
            "default": 5
          },
          "top_k_policies": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Top K Policies",
            "default": 4
          }
        },
        "type": "object",
        "required": [
          "issue_summary"
        ],
        "title": "ComplaintInvestigationRequest",
        "description": "Input request to investigate a customer complaint."
      },
...
      "CurrentUserResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "department": {
            "type": "string",
            "title": "Department"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "name",
          "role",
          "department"
        ],
        "title": "CurrentUserResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}